+++ ptsa/data/readers/binary.py
@@ -10,6 +10,11 @@ import struct
 class BinaryRawReader(BaseRawReader):
 
     def __init__(self,**kwargs):
+        if 'channels' in kwargs:
+            channels = kwargs['channels']
+            if channels.dtype.names is not None and 'channel_1' in channels.dtype.names:
+                raise IndexError
+
         super(BinaryRawReader, self).__init__(**kwargs)
         FileFormat = namedtuple('FileFormat', ['data_size', 'format_string'])
         self.file_format_dict = {
@@ -36,6 +41,7 @@ class BinaryRawReader(BaseRawReader):
         except KeyError:
             warnings.warn('Could not find data format definition in the params file. Will read the file assuming' \
                           ' data format is int16', RuntimeWarning)
+        self.channel_labels_to_string()
 
 
     def samplerate(self):
@@ -46,10 +52,10 @@ class BinaryRawReader(BaseRawReader):
         :return: {int} size of the files whose core name (dataroot) matches self.dataroot. Assumes ALL files with this
         dataroot are of the same length and uses first channel to determin the common file length
         """
-        if isinstance(self.channels[0], six.binary_type):
-            ch = self.channels[0].decode()
+        if isinstance(self.channel_labels[0], six.binary_type):
+            ch = self.channel_labels[0].decode()
         else:
-            ch = self.channels[0]
+            ch = self.channel_labels[0]
         eegfname = self.dataroot + '.' + ch
         return osp.getsize(eegfname)