+++ ptsa/data/readers/tal.py
@@ -4,6 +4,7 @@ import os
 import numpy as np
 import pandas as pd
 from ptsa.data.readers import BaseReader
+import warnings
 import traits.api
 
 __all__ = [
@@ -20,15 +21,20 @@ class TalReader(BaseReader,traits.api.HasTraits):
     filename = traits.api.Str
     struct_name = traits.api.Str
     struct_type = traits.api.Enum('bi','mono')
+    unpack = traits.api.Bool
 
-    def __init__(self, filename,struct_name='bpTalStruct',struct_type='bi'):
+    def __init__(self, filename,struct_name='bpTalStruct',struct_type='bi',unpack=True):
         """
         Keyword arguments
         -----------------
 
         :param filename {str} -  path to tal file or pairs.json file
-        :param struct_name {str} -  name of the struct to load
-        :param struct_type {str} - either 'mono', indicating a monopolar struct, or 'bi', indicating a bipolar struct
+        :param struct_name {str} -  name of the struct to load. Default is 'bpTalStruct'. Only relevant when reading
+        MATLAB tal struct files.
+        :param struct_type {str} - either 'mono', indicating a monopolar struct, or 'bi', indicating a bipolar struct.
+        Default is 'bi'.
+        :param unpack {bool} - If :py:val:False, returns the "atlases" column as an array of python dicts, rather
+        than unpacking them into nested structured arrays. Default is :py:True.
         :return: None
 
         """
@@ -40,11 +46,17 @@ class TalReader(BaseReader,traits.api.HasTraits):
         self.bipolar_channels=None
 
         self.tal_struct_array = None
+        self.unpack = unpack
+        if not self.unpack:
+            warnings.warn('Unpack option will be removed in a future release,'
+                          'at which point behavior will be as with unpack=True',
+                          PendingDeprecationWarning)
         self._json = os.path.splitext(self.filename)[-1]=='.json'
         if self.struct_type not in ['bi','mono']:
             raise AttributeError('Value %s not a valid struct_type. Please choose either "mono" or "bi"'%self.struct_type)
         if self.struct_type=='mono':
             self.struct_name='talStruct'
+
 
     def get_bipolar_pairs(self):
         """
@@ -80,25 +92,85 @@ class TalReader(BaseReader,traits.api.HasTraits):
         for i, channel_array in enumerate(channel_record_array):
             self.bipolar_channels[i] = tuple(map(lambda x: str(x).zfill(3), channel_array))
 
-    def from_dict(self,pairs):
-        keys = pairs.keys()
-        subject = [k for k in keys if k not in ['version','info','meta']][0]
+    @classmethod
+    def from_records(cls,contact_dict):
+        """
+        Helper method for :meth:from_dict.
+        Takes a list of records (dictionaries with semi-consistent fields)
+        and returns a structured array whose fields are the keys of each record.
+        Nested records are handled by recursion.
 
-        if self.struct_type=='bi':
-            pairs = pd.DataFrame.from_dict(pairs[subject]['pairs'], orient='index').sort_values(by=['channel_1','channel_2'])
-            pairs.index.name = 'tagName'
-            pairs['channel'] = [[ch1, ch2] for ch1, ch2 in zip(pairs.channel_1.values, pairs.channel_2.values)]
-            pairs['eType'] = pairs.type_1
-            return pairs.to_records()
-        elif self.struct_type == 'mono':
-            contacts = pd.DataFrame.from_dict(pairs[subject]['contacts'],orient='index').sort_values(by='channel')
-            contacts.index.name = 'tagName'
-            return contacts.to_records()
+        Missing entries should be represented by either None, NaN, or an empty dictionary.
+
+        :param contact_dict: {List[Union(Dict,None,NaN)]}
+        :return: {np.array} A structured array with the same indexing structure as :arg:contact_dict
+        """
+        contact_df = pd.DataFrame.from_records([x if not pd.isnull(x) else {} for x in contact_dict])
+        dict_cols = [col for col in contact_df.columns if any(isinstance(val,dict) for val in contact_df[col])]
+        flat_cols = [col for col in contact_df.columns if not col in dict_cols]
+        flat_df = contact_df[flat_cols]
+        dtype = cls.mkdtype(flat_df)
+
+        nested_arrs = [cls.from_records(contact_df[col]) for col in dict_cols]
+        nested_dtypes = [np.dtype([(str(col),x.dtype)]) for col,x in zip(dict_cols,nested_arrs)]
+        new_dtype = cls.merge_dtypes(dtype,*nested_dtypes)
+        new_arr = np.empty(len(contact_df),dtype=new_dtype)
+        for col in flat_cols:
+            new_arr[col] = flat_df[col].values
+        for (i,col) in enumerate(dict_cols):
+            new_arr[col] = nested_arrs[i]
+        return np.rec.array(new_arr)
+
+    def from_dict(self,json_dict,unpack=True):
+        """
+        Reads a JSON localization file into a record array.
+        :param json_dict: A dictionary of localization information, with one entry for each contact or pair.
+        :return:
+        """
+        keys = json_dict.keys()
+        subject = [k for k in keys if k not in ['version', 'info', 'meta']][0]
+        records =json_dict[subject]['contacts' if self.struct_type == 'mono' else 'pairs'].values()
+        if not unpack:
+            return pd.DataFrame.from_records(list(records)).to_records(index=False)
+        else:
+            ts = self.from_records(records)
+            if self.struct_type=='bi':
+                extended_dt = self.merge_dtypes(ts.dtype,np.dtype([('channel',int,2),
+                                                                   ('eType','U256'),('tagName','U256')]))
+            else:
+                extended_dt = self.merge_dtypes(ts.dtype,np.dtype([('eType','U2'),('tagName','U256')]))
+            # Add some additional fields for compatibility
+            new_ts = np.empty(ts.shape,dtype=extended_dt)
+            for col in ts.dtype.names:
+                new_ts[col] = ts[col]
+            if self.struct_type == 'bi':
+                new_ts['channel'][:,0] = ts['channel_1']
+                new_ts['channel'][:,1] = ts['channel_2']
+                new_ts['eType'] = ts['type_1']
+                new_ts['tagName'] = ts['code']
+            new_ts.sort(order='channel')
+            return np.rec.array(new_ts)
+
+
+    @classmethod
+    def mkdtype(cls,flat_df):
+        """
+        Invariant: there are no subfields of flat_df whose elements are dictionaries
+        :param flat_df: {pandas.DataFrame}
+        :return: {np.dtype}
+        """
+        dt_list =  [(str(c),flat_df[c].values.dtype if flat_df[c].values.dtype != np.dtype('O') else 'U256') for c in flat_df.columns]
+        return np.dtype(dt_list)
+
+    @classmethod
+    def merge_dtypes(cls,*dtypes):
+        if len(dtypes) == 1:
+            return dtypes[0]
+        elts = [(n,dt[n]) for dt in dtypes for n in dt.names]
+        return np.dtype(elts)
 
     def read(self):
-
         """
-
         :return: np.recarray representing tal struct array
         """
         if not self._json:
@@ -122,7 +194,7 @@ class TalReader(BaseReader,traits.api.HasTraits):
         else:
             with open(self.filename) as fp:
                 pairs= json.load(fp)
-            self.tal_struct_array = self.from_dict(pairs)
+            self.tal_struct_array = self.from_dict(pairs,unpack=self.unpack)
             return self.tal_struct_array
 
         raise AttributeError('Could not read tal struct data. Try specifying struct_name argument :'