+++ RELEASE_NOTES.rst
@@ -9,7 +9,7 @@ Summary of changes:
   - The 'atlases' field is a nested structured array, rather than an array of dictionaries
   - Most (?) idioms for getting coordinates and atlas labels for an electrode should continue to work in the new format
   - The 'channels' field has been changed from an object array of lists to an array of two integers
-
+* Added IPython notebook demonstrating reading localization information with TalReader (thanks @LoganJF !)
 
 
 Version 1.1.5